changing the existing tests and adding the light tests
