--- sep4_drivers/lib/sensor_readings/light_intensity.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-float calculate_light_intensity(uint16_t light_adc);
