--- sep4_drivers/lib/sensor_readings/moisture.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-int calculate_moisture_percentage(int sensor_value);
