+++ sep4_drivers/platformio.ini
@@ -8,6 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
+
 [env:win_test]
 platform = native
 build_flags = 
@@ -18,7 +19,10 @@ test_framework = unity
 lib_deps = throwtheswitch/Unity
 test_filter = test_moisture_percentage
 lib_ignore = MQTTPacket
-build_src_filter = +<test> -<*> 
+lib_compat_mode = off
+build_src_filter = 
+    +<test>
+    -<*>
 
 ; [env:target_test]
 ; platform = atmelavr
